+++ packages/nextra/src/client/hooks/index.ts
@@ -1,2 +1,3 @@
 export { useMounted } from './use-mounted.js'
 export { useFSRoute } from './use-fs-route.js'
+export { useHash } from './use-hash.js'