+++ packages/nextra/src/server/remark-plugins/remark-headings.ts
@@ -7,6 +7,7 @@ import type {
 } from 'mdast-util-mdx-jsx'
 import type { Plugin } from 'unified'
 import { visit } from 'unist-util-visit'
+import { visitChildren } from 'unist-util-visit-children'
 import type { Heading } from '../../types.js'
 import { MARKDOWN_EXTENSION_RE } from '../constants.js'
 import { createAstObject } from '../utils.js'
@@ -99,7 +100,11 @@ export const remarkHeadings: Plugin<
                       body: [
                         {
                           type: 'ExpressionStatement',
-                          expression: createAstObject({ display: 'none' })
+                          expression: createAstObject({
+                            visibility: 'hidden',
+                            width: 0,
+                            height: 0
+                          })
                         }
                       ]
                     }
@@ -108,6 +113,27 @@ export const remarkHeadings: Plugin<
               }
             ] satisfies MdxJsxAttribute[]
           } as any)
+        }
+
+        const isDetails =
+          node.type === 'mdxJsxFlowElement' && node.name === 'details'
+        if (isDetails) {
+          const visitor = visitChildren((node: any) => {
+            const isSummary =
+              node.type === 'mdxJsxTextElement' && node.name === 'summary'
+            if (isSummary) {
+              const value = getFlattenedValue(node)
+              const id = slugger.slug(value)
+              node.attributes.push({
+                type: 'mdxJsxAttribute',
+                name: 'id',
+                value: id
+              })
+            } else if ('children' in node) {
+              visitor(node)
+            }
+          })
+          visitor(node)
         }
 
         if (isRemoteContent) {